adding a more feature to scrolling to next slide
